requirements.txt
PyPDF2>=3.0.0
pdfminer.six>=20221105
tk>=0.1.0